Made it so the 'News and Alerts' side bar updater appears on all pages. Did this by adding a method in the application controller called 'sideposts_updates' Then applied this method to all controllers with a simple before_filter :sideposts_updates
Also started to add login_require filters to certain pages
These couple of methods are in the sessions_controller.rb
Added before_filter call to the following controllers :
documents
groups
users

You know have to be logged in to view anything in these controllers. If you try to access these controllers with out being logged in you will be thrown to the new session page.